EstruturasDeControle/bancoDeDados/requirements.txt
astroid==2.4.2
colorama==0.4.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector==2.2.9
pylint==2.6.0
six==1.15.0
toml==0.10.2
wrapt==1.12.1
